Adds the ability to specify a FailureRate on your endpoints
This will allow you to incorporate an element of chaos into your
tests and verify that your retry mechanisms are working as expected.

Adds some tests to try and ensure that there is an element of randomness.

If you're reading this, thoughts on how to better test probability reliably
would be appreciated!